Add authorization and authentication for hapi notes api be

+++ hapi-notes-backend/src/api/notes/handler.js
@@ -22,7 +22,10 @@ class NoteHandler {
   async postNoteHandler(request, h) {
     const { title = 'untitled', body, tags } = request.payload;
     this._validator.validateNotePayload(request.payload);
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: credentialId } = request.auth.credentials;
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -69,7 +72,8 @@ class NoteHandler {
   }
 
   async getNotesHandler(request, h) {
-    const notes = await this._service.getNotes();
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     const response = h.response({
       status: 'success',
       data: {
@@ -82,6 +86,8 @@ class NoteHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     const response = h.response({
       status: 'success',
@@ -128,6 +134,8 @@ class NoteHandler {
   async putNoteByIdHandler(request, h) {
     const { id } = request.params;
     this._validator.validateNotePayload(request.payload);
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
     await this._service.editNoteById(id, request.payload);
     const response = h.response({
       status: 'success',
@@ -168,6 +176,8 @@ class NoteHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
     await this._service.deleteNoteById(id);
     const response = h.response({
       status: 'success',
